// packages/core/src/generated/types.ts
export interface AdditionalDataType {
  /**
   * Unique identification of the data field
   *
   * Az adatmező egyedi azonosítója
   */
  dataName: string;
  /**
   * Description of content of the data field
   *
   * Az adatmező tartalmának szöveges leírása
   */
  dataDescription: string;
  /**
   * Value of the data
   *
   * Az adat értéke
   */
  dataValue: string;
}
// ...
export interface AdvanceDataType {
  /**
   * The value is true if the invoice item is a kind of advance charge
   *
   * Értéke true, ha a számla tétel előleg jellegű
   */
  advanceIndicator: boolean;
  /**
   * Advance payment related data
   *
   * Előleg fizetéshez kapcsolódó adatok
   */
  advancePaymentData?: AdvancePaymentDataType;
}

/**
 * Advance payment related data
 *
 * Előlegfizetéshez kapcsolódó adatok
 */
export interface AdvancePaymentDataType {
  /**
   * Invoice number containing the advance payment
   *
   * Az előlegszámlának a sorszáma, amelyben az előlegfizetés történt
   */
  advanceOriginalInvoice: string;
  /**
   * Payment date of the advance
   *
   * Az előleg fizetésének dátuma
   */
  advancePaymentDate: string;
  /**
   * Applied exchange rate of the advance
   *
   * Az előlegfizetés során alkalmazott árfolyam
   */
  advanceExchangeRate: string;
}
// ...
export interface ConventionalInvoiceInfoType {
  /**
   * Order numbers
   *
   * Megrendelésszám(ok)
   */
  orderNumbers?: OrderNumbersType;
  /**
   * Delivery notes
   *
   * Szállítólevél szám(ok)
   */
  deliveryNotes?: DeliveryNotesType;
  /**
   * Shipping dates
   *
   * Szállítási dátum(ok)
   */
  shippingDates?: ShippingDatesType;
  /**
   * Contract numbers
   *
   * Szerződésszám(ok)
   */
  contractNumbers?: ContractNumbersType;
  /**
   * Company codes of the supplier
   *
   * Az eladó vállalati kódja(i)
   */
  supplierCompanyCodes?: SupplierCompanyCodesType;
  /**
   * Company codes of the customer
   *
   * A vevő vállalati kódja(i)
   */
  customerCompanyCodes?: CustomerCompanyCodesType;
  /**
   * Dealer codes
   *
   * Beszállító kód(ok)
   */
  dealerCodes?: DealerCodesType;
  /**
   * Cost centers
   *
   * Költséghely(ek)
   */
  costCenters?: CostCentersType;
  /**
   * Project numbers
   *
   * Projektszám(ok)
   */
  projectNumbers?: ProjectNumbersType;
  /**
   * General ledger account numbers
   *
   * Főkönyvi számlaszám(ok)
   */
  generalLedgerAccountNumbers?: GeneralLedgerAccountNumbersType;
  /**
   * Supplier's global location numbers
   *
   * Kiállítói globális helyazonosító szám(ok)
   */
  glnNumbersSupplier?: GlnNumbersType;
  /**
   * Customer's global location numbers
   *
   * Vevői globális helyazonosító szám(ok)
   */
  glnNumbersCustomer?: GlnNumbersType;
  /**
   * Material numbers
   *
   * Anyagszám(ok)
   */
  materialNumbers?: MaterialNumbersType;
  /**
   * Item number(s)
   *
   * Cikkszám(ok)
   */
  itemNumbers?: ItemNumbersType;
  /**
   * EKAER ID-s
   *
   * EKÁER azonosító(k)
   */
  ekaerIds?: EkaerIdsType;
}

/**
 * Cost centers
 *
 * Költséghelyek
 */
export interface CostCentersType {
  /**
   * Cost center
   *
   * Költséghely
   */
  costCenter: string[];
}
// ...
export type InvoiceAppearanceType = 'PAPER' | 'ELECTRONIC' | 'EDI' | 'UNKNOWN';

/**
 * Type of invoice
 *
 * A számla típusa
 */
export type InvoiceCategoryType = 'NORMAL' | 'SIMPLIFIED' | 'AGGREGATE';
// ...
export interface InvoiceData {
  /**
   * Sequential number of the original invoice or modification document - section
   * 169 (b) or section 170 (1) b) of the VAT law
   *
   * Számla vagy módosító okirat sorszáma - ÁFA tv. 169. § b) vagy 170. § (1)
   * bek. b) pont
   */
  invoiceNumber: string;
  /**
   * Date of issue of the invoice or the modification document - section 169 (a)
   * of the VAT law, section 170 (1) a) of the VAT law
   *
   * Számla vagy módosító okirat kelte - ÁFA tv. 169. § a), ÁFA tv. 170. § (1)
   * bek. a)
   */
  invoiceIssueDate: string;
  /**
   * Indicates whether the data exchange is identical with the invoice (the
   * invoice does not contain any more data)
   *
   * Jelöli, ha az adatszolgáltatás maga a számla (a számlán nem szerepel több
   * adat)
   */
  completenessIndicator: boolean;
  /**
   * A common type to describe invoice information
   *
   * Számlaadatok leírására szolgáló közös típus
   */
  invoiceMain: InvoiceMainType;
}
// ...
export interface InvoiceDetailType {
  /**
   * Type of invoice. In case of modification document the type of original
   * invoice
   *
   * A számla típusa, módosító okirat esetén az eredeti számla típusa
   */
  invoiceCategory: InvoiceCategoryType;
  /**
   * Delivery date (if this field does not exist on the invoice, the date of the
   * invoice should be considered as such) - section 169 (g) of the VAT law
   *
   * Teljesítés dátuma (ha nem szerepel a számlán, akkor azonos a számla
   * keltével) - ÁFA tv. 169. § g)
   */
  invoiceDeliveryDate: string;
  /**
   * The first day of the delivery, if the invoice delivery is a period
   *
   * Amennyiben a számla egy időszakra vonatkozik, akkor az időszak első napja
   */
  invoiceDeliveryPeriodStart?: string;
  /**
   * The last day of the delivery, if the invoice delivery is a period
   *
   * Amennyiben a számla egy időszakra vonatkozik, akkor az időszak utolsó napja
   */
  invoiceDeliveryPeriodEnd?: string;
  /**
   * Date of accounting accomplishment. In the event of a period, the last day of
   * the period
   *
   * Számviteli teljesítés dátuma. Időszak esetén az időszak utolsó napja
   */
  invoiceAccountingDeliveryDate?: string;
  /**
   * Indicates where by agreement of the parties it gives rise to successive
   * statements of account or successive payments relating to the supply of
   * goods, or the supply of services, or if the consideration agreed upon for
   * such goods and/or services applies to specific periods.
   *
   * Annak jelzése, ha a felek a termékértékesítés, szolgáltatás nyújtás során
   * időszakonkénti elszámolásban vagy fizetésben állapodnak meg, vagy a
   * termékértékesítés, szolgáltatás nyújtás ellenértékét meghatározott időpontra
   * állapítják meg.
   */
  periodicalSettlement?: boolean;
  /**
   * Marking of low tax-bracket enterprise
   *
   * Kisadózó jelzése
   */
  smallBusinessIndicator?: boolean;
  /**
   * ISO 4217 currency code on the invoice
   *
   * A számla pénzneme az ISO 4217 szabvány szerint
   */
  currencyCode: string;
  /**
   * In case any currency is used other than HUF, the applied exchange rate
   * should be mentioned: 1 unit of the foreign currency expressed in HUF
   *
   * HUF-tól különböző pénznem esetén az alkalmazott árfolyam: egy egység értéke
   * HUF-ban
   */
  exchangeRate: string;
  /**
   * Marking the fact of utility settlement invoice (invoice according to Act
   * CLXXXVIII of 2013)
   *
   * Közmű elszámoló számla jelölése (2013.évi CLXXXVIII törvény szerinti
   * elszámoló számla)
   */
  utilitySettlementIndicator?: boolean;
  /**
   * Marking the fact of self-billing (in the case of self-billing the value is
   * true)
   *
   * Önszámlázás jelölése (önszámlázás esetén true)
   */
  selfBillingIndicator?: boolean;
  /**
   * Method of payment
   *
   * Fizetés módja
   */
  paymentMethod?: PaymentMethodType;
  /**
   * Deadline for payment
   *
   * Fizetési határidő
   */
  paymentDate?: string;
  /**
   * Marking the fact of cash accounting if this is indicated on the invoice -
   * section 169 (h) of the VAT law. The value is true in case of cash accounting
   *
   * Pénzforgalmi elszámolás jelölése, ha az szerepel a számlán - ÁFA tv. 169. §
   * h). Értéke true pénzforgalmi elszámolás esetén
   */
  cashAccountingIndicator?: boolean;
  /**
   * Form of appearance of the invoice or modification document
   *
   * A számla vagy módosító okirat megjelenési formája
   */
  invoiceAppearance: InvoiceAppearanceType;
  /**
   * Other conventionally named data to assist in invoice processing
   *
   * A számlafeldolgozást segítő, egyezményesen nevesített egyéb adatok
   */
  conventionalInvoiceInfo?: ConventionalInvoiceInfoType;
  /**
   * Other data in relation to the invoice
   *
   * A számlára vonatkozó egyéb adat
   */
  additionalInvoiceData?: AdditionalDataType[];
}
// ...
export interface InvoiceHeadType {
  /**
   * Data related to the issuer of the invoice (supplier)
   *
   * Számla kibocsátó (eladó) adatai
   */
  supplierInfo: SupplierInfoType;
  /**
   * Data related to the customer
   *
   * Vevő adatai
   */
  customerInfo?: CustomerInfoType;
  /**
   * Data related to the fiscal representative
   *
   * Pénzügyi képviselő adatai
   */
  fiscalRepresentativeInfo?: FiscalRepresentativeType;
  /**
   * Invoice detail adata
   *
   * Számla részletező adatok
   */
  invoiceDetail: InvoiceDetailType;
}
// ...
export interface InvoiceMainType {
  /**
   * Data of a single invoice or modification document
   *
   * Egy számla vagy módosító okirat adatai
   */
  invoice?: InvoiceType;
  /**
   * Data of a batch of modification documents
   *
   * Kötegelt módosító okirat adatai
   */
  batchInvoice?: BatchInvoiceType[];
}
// ...
export interface InvoiceReferenceType {
  /**
   * Sequence number of the original invoice, on which the modification occurs -
   * section 170 (1) c) of the VAT law
   *
   * Az eredeti számla sorszáma, melyre a módosítás vonatkozik - ÁFA tv. 170. §
   * (1) c)
   */
  originalInvoiceNumber: string;
  /**
   * Indicates whether the modification references to an original invoice which
   * is not and will not be exchanged
   *
   * Annak jelzése, hogy a módosítás olyan alapszámlára hivatkozik, amelyről nem
   * történt és nem is fog történni adatszolgáltatás
   */
  modifyWithoutMaster: boolean;
  /**
   * The unique sequence number referring to the original invoice
   *
   * A számlára vonatkozó módosító okirat egyedi sorszáma
   */
  modificationIndex: number;
}

/**
 * Processing status of the invoice
 *
 * A számla feldolgozási státusza
 */
export type InvoiceStatusType = 'RECEIVED' | 'PROCESSING' | 'SAVED' | 'DONE' | 'ABORTED';

/**
 * Data of a single invoice or modification document
 *
 * Egy számla vagy módosító okirat adatai
 */
export interface InvoiceType {
  /**
   * Modification or cancellation data
   *
   * A módosítás vagy érvénytelenítés adatai
   */
  invoiceReference?: InvoiceReferenceType;
  /**
   * Data concerning the whole invoice
   *
   * A számla egészét jellemző adatok
   */
  invoiceHead: InvoiceHeadType;
  /**
   * Product/service data appearing on the invoice
   *
   * A számlán szereplő tételek adatai
   */
  invoiceLines?: LinesType;
  /**
   * Summary data of product charges
   *
   * Termékdíjjal kapcsolatos összesítő adatok
   *
   * @remarks At most 2 entries.
   */
  productFeeSummary?: ProductFeeSummaryType[];
  /**
   * Summary data according to VAT law
   *
   * Az ÁFA törvény szerinti összesítő adatok
   */
  invoiceSummary: SummaryType;
}
// ...
export interface LinesType {
  /**
   * Indicates whether the data exchange contains merged line data due to size
   * reduction
   *
   * Jelöli, ha az adatszolgáltatás méretcsökkentés miatt összevont soradatokat
   * tartalmaz
   */
  mergedItemIndicator: boolean;
  /**
   * Product / service item
   *
   * Termék/szolgáltatás tétel
   */
  line: LineType[];
}
// ...
export type PaymentMethodType = 'TRANSFER' | 'CASH' | 'CARD' | 'VOUCHER' | 'OTHER';
// ...
export interface ProductFeeSummaryType {
  /**
   * Indicating whether the the product fee summary concerns refund or deposit
   *
   * Annak jelzése, hogy a termékdíj összesítés visszaigénylésre (REFUND) vagy
   * raktárba történő beszállításra (DEPOSIT) vonatkozik
   */
  productFeeOperation: ProductFeeOperationType;
  /**
   * Product charges data
   *
   * Termékdíj adatok
   */
  productFeeData: ProductFeeDataType[];
  /**
   * Aggregate product charges
   *
   * Termékdíj összesen
   */
  productChargeSum: string;
  /**
   * Data of the document verifying the declaration submitted on the product fee
   * according to the Paragraph (3), Section 13 and the Paragraph (3) Section 25
   * of the Act LXXXV of 2011
   *
   * A termékdíj bevallását igazoló dokumentum adatai a 2011. évi LXXXV. tv. 13.
   * § (3) szerint és a 25. § (3) szerint
   */
  paymentEvidenceDocumentData?: PaymentEvidenceDocumentDataType;
}
// ...
export type UnitOfMeasureType =
  | 'PIECE'
  | 'KILOGRAM'
  | 'TON'
  | 'KWH'
  | 'DAY'
  | 'HOUR'
  | 'MINUTE'
  | 'MONTH'
  | 'LITER'
  | 'KILOMETER'
  | 'CUBIC_METER'
  | 'METER'
  | 'LINEAR_METER'
  | 'CARTON'
  | 'PACK'
  | 'OWN';